# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Ultralytics Analytics API
  description: |
    Self-hosted analytics platform for tracking user events and page views.
    
    ## Authentication
    All API endpoints (except `/health`) require an API key passed via the `X-API-Key` header.
  version: 0.4.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    email: bubba@thegivingchain.com

servers:
  - url: http://localhost:3000
    description: Local development server

tags:
  - name: Health
    description: Health check endpoints
  - name: Events
    description: Event tracking and retrieval

paths:
  /health:
    get:
      tags:
        - Health
      summary: Health check
      description: Returns server and database health status
      responses:
        '200':
          description: Server is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /api/events:
    get:
      tags:
        - Events
      summary: Query events
      description: Retrieve events with optional date range filtering
      security:
        - ApiKeyAuth: []
      parameters:
        - name: start_date
          in: query
          description: Start date for filtering (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          description: End date for filtering (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          description: Maximum number of events to return
          schema:
            type: integer
            default: 100
            maximum: 1000
        - name: offset
          in: query
          description: Number of events to skip
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of events
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
                  total:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Events
      summary: Track an event
      description: Record a new analytics event
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventInput'
      responses:
        '201':
          description: Event recorded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/events/batch:
    post:
      tags:
        - Events
      summary: Track multiple events
      description: Record multiple analytics events in a single request
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                events:
                  type: array
                  items:
                    $ref: '#/components/schemas/EventInput'
                  maxItems: 1000
              required:
                - events
      responses:
        '201':
          description: Events recorded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  processed:
                    type: integer
                  failed:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key

  schemas:
    Event:
      type: object
      properties:
        id:
          type: string
          format: uuid
        event_type:
          type: string
        event_name:
          type: string
        properties:
          type: object
        session_id:
          type: string
        user_id:
          type: string
        timestamp:
          type: string
          format: date-time

    EventInput:
      type: object
      properties:
        event_type:
          type: string
          enum: [pageview, event, identify]
        event_name:
          type: string
        properties:
          type: object
        session_id:
          type: string
        user_id:
          type: string
      required:
        - event_type

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          enum: [healthy, unhealthy]
        database:
          type: string
          enum: [connected, disconnected]
        uptime:
          type: number
        timestamp:
          type: string
          format: date-time

    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'